Move summary related components to the summary-section folder
