+++ js/src/dashboard/summary-section/index.scss
