+++ js/src/dashboard/summary-section/paid-campaign-promotion-card.scss
